fix: add makeBuildOptionsDefaulted
